+++ src/components/select/simple-select/index.ts
@@ -1,2 +1,5 @@
 export { default } from "./simple-select.component";
-export type { SimpleSelectProps } from "./simple-select.component";
+export type {
+  SimpleSelectProps,
+  CustomSelectChangeEvent,
+} from "./simple-select.component";